Drop support for MySQL 5.6 and older and MariaDB 10.2.6 and older
