+++ .github/workflows/continuous-integration.yml
@@ -225,7 +225,6 @@ jobs:
         php-version:
           - "8.0"
         mariadb-version:
-          - "10.0"
           - "10.2"
           - "10.5"
         extension:
