+++ docs/en/reference/platforms.rst
@@ -33,14 +33,13 @@ can be found as follows:
 MySQL
 ^^^^^
 
--  ``MySQLPlatform`` for version 5.0 and above.
--  ``MySQL57Platform`` for version 5.7 (5.7.9 GA) and above.
+-  ``MySQLPlatform`` for version 5.7 (5.7.9 GA) and above.
 -  ``MySQL80Platform`` for version 8.0 (8.0 GA) and above.
 
 MariaDB
 ^^^^^
 
--  ``MariaDb1027Platform`` for version 10.2 (10.2.7 GA) and above.
+-  ``MariaDBPlatform`` for version 10.2 (10.2.7 GA) and above.
 
 Oracle
 ^^^^^^
