+++ psalm.xml.dist
@@ -39,17 +39,6 @@
                 <file name="src/Driver/OCI8/ConvertPositionalToNamedPlaceholders.php"/>
             </errorLevel>
         </ConflictingReferenceConstraint>
-        <DeprecatedClass>
-            <errorLevel type="suppress">
-                <!--
-                    TODO: remove in 4.0.0
-                -->
-                <referencedClass name="Doctrine\DBAL\Platforms\Keywords\MariaDb102Keywords"/>
-                <referencedClass name="Doctrine\DBAL\Platforms\Keywords\MySQL57Keywords"/>
-                <referencedClass name="Doctrine\DBAL\Platforms\MariaDb1027Platform"/>
-                <referencedClass name="Doctrine\DBAL\Platforms\MySQL57Platform"/>
-            </errorLevel>
-        </DeprecatedClass>
         <DeprecatedMethod>
             <errorLevel type="suppress">
                 <!--
