+++ src/Driver/AbstractMySQLDriver.php
@@ -11,13 +11,11 @@ use Doctrine\DBAL\Exception;
 use Doctrine\DBAL\Platforms\AbstractMySQLPlatform;
 use Doctrine\DBAL\Platforms\AbstractPlatform;
 use Doctrine\DBAL\Platforms\Exception\InvalidPlatformVersion;
-use Doctrine\DBAL\Platforms\MariaDb1027Platform;
-use Doctrine\DBAL\Platforms\MySQL57Platform;
+use Doctrine\DBAL\Platforms\MariaDBPlatform;
 use Doctrine\DBAL\Platforms\MySQL80Platform;
 use Doctrine\DBAL\Platforms\MySQLPlatform;
 use Doctrine\DBAL\Schema\MySQLSchemaManager;
 use Doctrine\DBAL\ServerVersionProvider;
-use Doctrine\Deprecations\Deprecation;
 
 use function assert;
 use function preg_match;
@@ -34,31 +32,16 @@ abstract class AbstractMySQLDriver implements Driver
      *
      * @throws Exception
      */
-    public function getDatabasePlatform(ServerVersionProvider $versionProvider): MySQLPlatform
+    public function getDatabasePlatform(ServerVersionProvider $versionProvider): AbstractMySQLPlatform
     {
         $version = $versionProvider->getServerVersion();
-        $mariadb = stripos($version, 'mariadb') !== false;
-        if ($mariadb && version_compare($this->getMariaDbMysqlVersionNumber($version), '10.2.7', '>=')) {
-            return new MariaDb1027Platform();
+        if (stripos($version, 'mariadb') !== false) {
+            return new MariaDBPlatform();
         }
 
-        if (! $mariadb) {
-            $oracleMysqlVersion = $this->getOracleMysqlVersionNumber($version);
-            if (version_compare($oracleMysqlVersion, '8', '>=')) {
-                return new MySQL80Platform();
-            }
-
-            if (version_compare($oracleMysqlVersion, '5.7.9', '>=')) {
-                return new MySQL57Platform();
-            }
+        if (version_compare($this->getOracleMysqlVersionNumber($version), '8', '>=')) {
+            return new MySQL80Platform();
         }
-
-        Deprecation::trigger(
-            'doctrine/dbal',
-            'https://github.com/doctrine/dbal/pull/5060',
-            'MySQL 5.6 support is deprecated and will be removed in DBAL 4.'
-                . ' Consider upgrading to MySQL 5.7 or later.'
-        );
 
         return new MySQLPlatform();
     }
@@ -95,32 +78,6 @@ abstract class AbstractMySQLDriver implements Driver
         }
 
         return $majorVersion . '.' . $minorVersion . '.' . $patchVersion;
-    }
-
-    /**
-     * Detect MariaDB server version, including hack for some mariadb distributions
-     * that starts with the prefix '5.5.5-'
-     *
-     * @param string $versionString Version string as returned by mariadb server, i.e. '5.5.5-Mariadb-10.0.8-xenial'
-     *
-     * @throws Exception
-     */
-    private function getMariaDbMysqlVersionNumber(string $versionString): string
-    {
-        if (
-            preg_match(
-                '/^(?:5\.5\.5-)?(mariadb-)?(?P<major>\d+)\.(?P<minor>\d+)\.(?P<patch>\d+)/i',
-                $versionString,
-                $versionParts
-            ) === 0
-        ) {
-            throw InvalidPlatformVersion::new(
-                $versionString,
-                '^(?:5\.5\.5-)?(mariadb-)?<major_version>.<minor_version>.<patch_version>'
-            );
-        }
-
-        return $versionParts['major'] . '.' . $versionParts['minor'] . '.' . $versionParts['patch'];
     }
 
     public function getSchemaManager(Connection $conn, AbstractPlatform $platform): MySQLSchemaManager
