--- src/Platforms/Keywords/MariaDb102Keywords.php
@@ -1,20 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Doctrine\DBAL\Platforms\Keywords;
-
-/**
- * MariaDb reserved keywords list.
- *
- * @deprecated Use {@link MariaDBKeywords} instead.
- *
- * @link https://mariadb.com/kb/en/the-mariadb-library/reserved-words/
- */
-final class MariaDb102Keywords extends MariaDBKeywords
-{
-    public function getName(): string
-    {
-        return 'MariaDb102';
-    }
-}
