+++ src/Platforms/Keywords/MySQLKeywords.php
@@ -16,6 +16,8 @@ class MySQLKeywords extends KeywordList
 
     /**
      * {@inheritdoc}
+     *
+     * @link https://dev.mysql.com/doc/mysqld-version-reference/en/keywords-5-7.html
      */
     protected function getKeywords(): array
     {
@@ -46,7 +48,6 @@ class MySQLKeywords extends KeywordList
             'COLLATE',
             'COLUMN',
             'CONDITION',
-            'CONNECTION',
             'CONSTRAINT',
             'CONTINUE',
             'CONVERT',
@@ -96,8 +97,8 @@ class MySQLKeywords extends KeywordList
             'FOREIGN',
             'FROM',
             'FULLTEXT',
-            'GENERAL',
-            'GOTO',
+            'GENERATED',
+            'GET',
             'GRANT',
             'GROUP',
             'HAVING',
@@ -107,7 +108,6 @@ class MySQLKeywords extends KeywordList
             'HOUR_SECOND',
             'IF',
             'IGNORE',
-            'IGNORE_SERVER_IDS',
             'IN',
             'INDEX',
             'INFILE',
@@ -124,13 +124,14 @@ class MySQLKeywords extends KeywordList
             'INTEGER',
             'INTERVAL',
             'INTO',
+            'IO_AFTER_GTIDS',
+            'IO_BEFORE_GTIDS',
             'IS',
             'ITERATE',
             'JOIN',
             'KEY',
             'KEYS',
             'KILL',
-            'LABEL',
             'LEADING',
             'LEAVE',
             'LEFT',
@@ -147,7 +148,7 @@ class MySQLKeywords extends KeywordList
             'LONGTEXT',
             'LOOP',
             'LOW_PRIORITY',
-            'MASTER_HEARTBEAT_PERIOD',
+            'MASTER_BIND',
             'MASTER_SSL_VERIFY_SERVER_CERT',
             'MATCH',
             'MAXVALUE',
@@ -166,6 +167,7 @@ class MySQLKeywords extends KeywordList
             'NUMERIC',
             'ON',
             'OPTIMIZE',
+            'OPTIMIZER_COSTS',
             'OPTION',
             'OPTIONALLY',
             'OR',
@@ -178,13 +180,11 @@ class MySQLKeywords extends KeywordList
             'PRIMARY',
             'PROCEDURE',
             'PURGE',
-            'RAID0',
             'RANGE',
             'READ',
             'READ_WRITE',
             'READS',
             'REAL',
-            'RECURSIVE',
             'REFERENCES',
             'REGEXP',
             'RELEASE',
@@ -198,7 +198,6 @@ class MySQLKeywords extends KeywordList
             'REVOKE',
             'RIGHT',
             'RLIKE',
-            'ROWS',
             'SCHEMA',
             'SCHEMAS',
             'SECOND_MICROSECOND',
@@ -208,9 +207,7 @@ class MySQLKeywords extends KeywordList
             'SET',
             'SHOW',
             'SIGNAL',
-            'SLOW',
             'SMALLINT',
-            'SONAME',
             'SPATIAL',
             'SPECIFIC',
             'SQL',
@@ -222,6 +219,7 @@ class MySQLKeywords extends KeywordList
             'SQLWARNING',
             'SSL',
             'STARTING',
+            'STORED',
             'STRAIGHT_JOIN',
             'TABLE',
             'TERMINATED',
@@ -250,12 +248,12 @@ class MySQLKeywords extends KeywordList
             'VARCHAR',
             'VARCHARACTER',
             'VARYING',
+            'VIRTUAL',
             'WHEN',
             'WHERE',
             'WHILE',
             'WITH',
             'WRITE',
-            'X509',
             'XOR',
             'YEAR_MONTH',
             'ZEROFILL',
