--- src/Platforms/MySQL57Platform.php
@@ -1,75 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Doctrine\DBAL\Platforms;
-
-use Doctrine\DBAL\Platforms\Keywords\KeywordList;
-use Doctrine\DBAL\Platforms\Keywords\MySQL57Keywords;
-use Doctrine\DBAL\Schema\Index;
-use Doctrine\DBAL\Schema\TableDiff;
-use Doctrine\DBAL\SQL\Parser;
-use Doctrine\DBAL\Types\Types;
-
-/**
- * Provides the behavior, features and SQL dialect of the MySQL 5.7 (5.7.9 GA) database platform.
- *
- * @deprecated This class will be merged with {@see MySQLPlatform} in 4.0 because support for MySQL
- *             releases prior to 5.7 will be dropped.
- */
-class MySQL57Platform extends MySQLPlatform
-{
-    public function hasNativeJsonType(): bool
-    {
-        return true;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getJsonTypeDeclarationSQL(array $column): string
-    {
-        return 'JSON';
-    }
-
-    public function createSQLParser(): Parser
-    {
-        return new Parser(true);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    protected function getPreAlterTableRenameIndexForeignKeySQL(TableDiff $diff): array
-    {
-        return [];
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    protected function getPostAlterTableRenameIndexForeignKeySQL(TableDiff $diff): array
-    {
-        return [];
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    protected function getRenameIndexSQL(string $oldIndexName, Index $index, string $tableName): array
-    {
-        return ['ALTER TABLE ' . $tableName . ' RENAME INDEX ' . $oldIndexName . ' TO ' . $index->getQuotedName($this)];
-    }
-
-    protected function createReservedKeywordsList(): KeywordList
-    {
-        return new MySQL57Keywords();
-    }
-
-    protected function initializeDoctrineTypeMappings(): void
-    {
-        parent::initializeDoctrineTypeMappings();
-
-        $this->doctrineTypeMapping['json'] = Types::JSON;
-    }
-}
