+++ tests/Platforms/MySQLPlatformTest.php
@@ -5,10 +5,9 @@ declare(strict_types=1);
 namespace Doctrine\DBAL\Tests\Platforms;
 
 use Doctrine\DBAL\Platforms\AbstractPlatform;
-use Doctrine\DBAL\Platforms\MySQL;
 use Doctrine\DBAL\Platforms\MySQLPlatform;
-use Doctrine\DBAL\Schema\Comparator;
 use Doctrine\DBAL\TransactionIsolationLevel;
+use Doctrine\DBAL\Types\Types;
 
 class MySQLPlatformTest extends AbstractMySQLPlatformTestCase
 {
@@ -17,9 +16,66 @@ class MySQLPlatformTest extends AbstractMySQLPlatformTestCase
         return new MySQLPlatform();
     }
 
-    protected function createComparator(): Comparator
+    public function testHasNativeJsonType(): void
     {
-        return new MySQL\Comparator($this->platform);
+        self::assertTrue($this->platform->hasNativeJsonType());
+    }
+
+    public function testReturnsJsonTypeDeclarationSQL(): void
+    {
+        self::assertSame('JSON', $this->platform->getJsonTypeDeclarationSQL([]));
+    }
+
+    public function testInitializesJsonTypeMapping(): void
+    {
+        self::assertTrue($this->platform->hasDoctrineTypeMappingFor('json'));
+        self::assertSame(Types::JSON, $this->platform->getDoctrineTypeMapping('json'));
+    }
+
+    /**
+     * @return string[]
+     */
+    protected function getAlterTableRenameIndexSQL(): array
+    {
+        return ['ALTER TABLE mytable RENAME INDEX idx_foo TO idx_bar'];
+    }
+
+    /**
+     * @return string[]
+     */
+    protected function getQuotedAlterTableRenameIndexSQL(): array
+    {
+        return [
+            'ALTER TABLE `table` RENAME INDEX `create` TO `select`',
+            'ALTER TABLE `table` RENAME INDEX `foo` TO `bar`',
+        ];
+    }
+
+    /**
+     * @return string[]
+     */
+    protected function getAlterTableRenameIndexInSchemaSQL(): array
+    {
+        return ['ALTER TABLE myschema.mytable RENAME INDEX idx_foo TO idx_bar'];
+    }
+
+    /**
+     * @return string[]
+     */
+    protected function getQuotedAlterTableRenameIndexInSchemaSQL(): array
+    {
+        return [
+            'ALTER TABLE `schema`.`table` RENAME INDEX `create` TO `select`',
+            'ALTER TABLE `schema`.`table` RENAME INDEX `foo` TO `bar`',
+        ];
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    protected function getGeneratesAlterTableRenameIndexUsedByForeignKeySQL(): array
+    {
+        return ['ALTER TABLE mytable RENAME INDEX idx_foo TO idx_foo_renamed'];
     }
 
     public function testHasCorrectDefaultTransactionIsolationLevel(): void
